==> 1_Data_Extraction.ipynb <==
### Import Modules

import pandas as pd
from bs4 import BeautifulSoup
import requests

### Manchester City Players Stats

mc_stats_url = "https://www.espn.in/football/team/stats/_/id/382/league/UEFA.CHAMPIONS/view/scoring"
mc_stats_response = requests.get(mc_stats_url)

mc_stats_soup = BeautifulSoup(mc_stats_response.content, 'html.parser')
mc_stats_table = mc_stats_soup.find("table")

mc_stats_headers = [th.text for th in mc_stats_table.find_all('th')]
mc_stats_headers

for i in range(len(mc_stats_headers)):
    if mc_stats_headers[i] == 'P':
        mc_stats_headers[i] = 'matches_played'
    if mc_stats_headers[i] == 'G':
        mc_stats_headers[i] = 'goals_scored'

mc_stats_data = []

rows = mc_stats_table.find_all('tr')
for row in rows:
    if(row.find('th')):
        continue
    row_data = [td.text for td in row.find_all('td')]
    mc_stats_data.append(row_data)

mc_stats_df = pd.DataFrame(mc_stats_data, columns=mc_stats_headers)
mc_stats_df.head()

mc_stats_df.drop(['RK'], axis=1, inplace=True)
mc_stats_df.head()

mc_stats_df.to_csv('./Data/ManchesterCityPlayersStats.csv', index=False)

### Manchester City Players Discipline

mc_disc_url = "https://www.espn.in/football/team/stats/_/id/382/league/UEFA.CHAMPIONS/view/discipline"
mc_disc_response = requests.get(mc_disc_url)

mc_disc_soup = BeautifulSoup(mc_disc_response.content, 'html.parser')
mc_disc_table = mc_disc_soup.find("table")

mc_disc_headers = [th.text for th in mc_disc_table.find_all('th')]
mc_disc_headers

for i in range(len(mc_disc_headers)):
    if mc_disc_headers[i] == 'P':
        mc_disc_headers[i] = 'matches_played'
    if mc_disc_headers[i] == 'yc':
        mc_disc_headers[i] = 'yellow_cards'
    if mc_disc_headers[i] == 'rc':
        mc_disc_headers[i] = 'red_cards'
    if mc_disc_headers[i] == 'Pts':
        mc_disc_headers[i] = 'points'

mc_disc_data = []

rows = mc_disc_table.find_all('tr')
for row in rows:
    if(row.find('th')):
        continue
    row_data = [td.text for td in row.find_all('td')]
    mc_disc_data.append(row_data)

mc_disc_df = pd.DataFrame(mc_disc_data, columns=mc_disc_headers)
mc_disc_df.head()

mc_disc_df.drop(['RK'], axis=1, inplace=True)
mc_disc_df.head()

mc_disc_df.to_csv('./Data/ManchesterCityPlayersDisc.csv', index=False)

### Inter Milan Players Stats

im_stats_url = "https://www.espn.in/football/team/stats/_/id/110/league/UEFA.CHAMPIONS/view/scoring"
im_stats_response = requests.get(im_stats_url)

im_stats_soup = BeautifulSoup(im_stats_response.content, 'html.parser')
im_stats_table = im_stats_soup.find("table")

im_stats_headers = [th.text for th in im_stats_table.find_all('th')]
im_stats_headers

for i in range(len(im_stats_headers)):
    if im_stats_headers[i] == 'P':
        im_stats_headers[i] = 'matches_played'
    if im_stats_headers[i] == 'G':
        im_stats_headers[i] = 'goals_scored'

im_stats_data = []

rows = im_stats_table.find_all('tr')
for row in rows:
    if(row.find('th')):
        continue
    row_data = [td.text for td in row.find_all('td')]
    im_stats_data.append(row_data)

im_stats_df = pd.DataFrame(im_stats_data, columns=im_stats_headers)
im_stats_df.head()

im_stats_df.drop(['RK'], axis=1, inplace=True)
im_stats_df.head()

im_stats_df.to_csv('./Data/InterMilanPlayersStats.csv', index=False)

### Inter Milan Players Discipline

im_disc_url = "https://www.espn.in/football/team/stats/_/id/110/league/UEFA.CHAMPIONS/view/discipline"
im_disc_response = requests.get(im_disc_url)

im_disc_soup = BeautifulSoup(im_disc_response.content, 'html.parser')
im_disc_table = im_disc_soup.find("table")

im_disc_headers = [th.text for th in im_disc_table.find_all('th')]
im_disc_headers

for i in range(len(im_disc_headers)):
    if im_disc_headers[i] == 'P':
        im_disc_headers[i] = 'matches_played'
    if im_disc_headers[i] == 'yc':
        im_disc_headers[i] = 'yellow_cards'
    if im_disc_headers[i] == 'rc':
        im_disc_headers[i] = 'red_cards'
    if im_disc_headers[i] == 'Pts':
        im_disc_headers[i] = 'points'

im_disc_data = []

rows = im_disc_table.find_all('tr')
for row in rows:
    if(row.find('th')):
        continue
    row_data = [td.text for td in row.find_all('td')]
    im_disc_data.append(row_data)

im_disc_df = pd.DataFrame(im_disc_data, columns=im_disc_headers)
im_disc_df.head()

im_disc_df.drop(['RK'], axis=1, inplace=True)
im_disc_df.head()

im_disc_df.to_csv('./Data/InterMilanPlayersDisc.csv', index=False)

### Merging Data

mc_stats = './Data/ManchesterCityPlayersStats.csv'
mc_disc = './Data/ManchesterCityPlayersDisc.csv'

mc_matches_goals_data = pd.read_csv(mc_stats)
mc_discipline_points_data = pd.read_csv(mc_disc)

mc_merged_data = pd.merge(mc_matches_goals_data, mc_discipline_points_data, on='Name', how='left').fillna(0)
mc_data = './Data/ManchesterCity.csv'
mc_merged_data.to_csv(mc_data, index=False)

mc_df = pd.read_csv(mc_data)
mc_df.drop(['matches_played_y'], axis=1, inplace=True)

mc_df.insert(1, "Team", "Manchester City", True)
mc_df.rename(columns = {'matches_played_x':'matches_played'}, inplace = True)
mc_df.head()

mc_df.to_csv(mc_data, index=False)

im_stats = './Data/InterMilanPlayersStats.csv'
im_disc = './Data/InterMilanPlayersDisc.csv'

im_matches_goals_data = pd.read_csv(im_stats)
im_discipline_points_data = pd.read_csv(im_disc)

im_merged_data = pd.merge(im_matches_goals_data, im_discipline_points_data, on='Name')
im_data = './Data/InterMilan.csv'
im_merged_data.to_csv(im_data, index=False)

im_df = pd.read_csv(im_data)
im_df.drop(['matches_played_y'], axis=1, inplace=True)

im_df.insert(1, "Team", "Inter Milan", True)
im_df.rename(columns = {'matches_played_x':'matches_played'}, inplace = True)
im_df.head()

im_df.to_csv(im_data, index=False)
