==> 2_Assigning_Weights_to_Factors.ipynb <==
### Import Modules and Dataset

import pandas as pd

man_city_data = pd.read_csv('./Data/ManchesterCity.csv')
inter_milan_data = pd.read_csv('./Data/InterMilan.csv')

man_city_data.head()

inter_milan_data.head()

man_city_data['normalized_matches_played'] = man_city_data['matches_played'] / man_city_data['matches_played'].max()
man_city_data['normalized_goals_scored'] = man_city_data['goals_scored'] / man_city_data['goals_scored'].max()
man_city_data['normalized_yellow_cards'] = man_city_data['yellow_cards'] / man_city_data['yellow_cards'].max()
man_city_data['normalized_red_cards'] = man_city_data['red_cards'] / man_city_data['red_cards'].max()
man_city_data['normalized_points'] = man_city_data['points'] / man_city_data['points'].max()

man_city_data.head()

inter_milan_data['normalized_matches_played'] = inter_milan_data['matches_played'] / inter_milan_data['matches_played'].max()
inter_milan_data['normalized_goals_scored'] = inter_milan_data['goals_scored'] / inter_milan_data['goals_scored'].max()
inter_milan_data['normalized_yellow_cards'] = inter_milan_data['yellow_cards'] / inter_milan_data['yellow_cards'].max()
inter_milan_data['normalized_red_cards'] = inter_milan_data['red_cards'] / inter_milan_data['red_cards'].max()
inter_milan_data['normalized_points'] = inter_milan_data['points'] / inter_milan_data['points'].max()

inter_milan_data.head()

inter_milan_data['normalized_red_cards'] = inter_milan_data['normalized_red_cards'].fillna(0)

inter_milan_data.head()

weights = {
    'Matches Played': 0.2,
    'Goals Scored': 0.4,
    'Yellow Cards': 0.1,
    'Red Cards': 0.1,
    'Points': 0.2
}

# Calculate the weighted scores for Manchester City
man_city_data['Weighted_Score'] = man_city_data['normalized_matches_played'] * weights['Matches Played'] \
                                 + man_city_data['normalized_goals_scored'] * weights['Goals Scored'] \
                                 - man_city_data['normalized_yellow_cards'] * weights['Yellow Cards'] \
                                 - man_city_data['normalized_red_cards'] * weights['Red Cards'] \
                                 - man_city_data['normalized_points'] * weights['Points']

man_city_data.head()

inter_milan_data['Weighted_Score'] = inter_milan_data['normalized_matches_played'] * weights['Matches Played'] \
                                     + inter_milan_data['normalized_goals_scored'] * weights['Goals Scored'] \
                                     - inter_milan_data['normalized_yellow_cards'] * weights['Yellow Cards'] \
                                     - inter_milan_data['normalized_red_cards'] * weights['Red Cards'] \
                                     - inter_milan_data['normalized_points'] * weights['Points']

inter_milan_data.head()

combined_data = pd.concat([man_city_data, inter_milan_data], ignore_index=True)

combined_data.head()

combined_data = combined_data.sort_values(['Team', 'Weighted_Score'], ascending=[True, False])

combined_data.head()

output_file = './Data/all_players_stats.csv'
combined_data.to_csv(output_file, index=False)
