==> 3_Final_Team_Selection.ipynb <==
import pandas as pd
combined_data = pd.read_csv('./Data/all_players_stats.csv')

combined_data.head()

sorted_players = combined_data.sort_values('Weighted_Score', ascending=False)
num_players_per_team = 11

man_city_players = sorted_players[sorted_players['Team'] == 'Manchester City'].head(num_players_per_team)
inter_milan_players = sorted_players[sorted_players['Team'] == 'Inter Milan'].head(num_players_per_team)

print("Manchester City Team Selection:")
print(man_city_players[['Name', 'Weighted_Score']])
print()
print("Inter Milan Team Selection:")
print(inter_milan_players[['Name', 'Weighted_Score']])

Manchester_City_Players_Weight = man_city_players['Weighted_Score'].sum()
Inter_Milan_Players_Weight = inter_milan_players['Weighted_Score'].sum()

print("Manchester City Players points: ", Manchester_City_Players_Weight)
print("Inter Milan Players points: ", Inter_Milan_Players_Weight)

all_players = pd.concat([man_city_players, inter_milan_players], axis=0)
all_players.sort_values(['Weighted_Score'], ascending=False, inplace=True)

all_players

final_team = all_players.head(11)
final_team[['Name', 'Team', 'matches_played', 'goals_scored', 'Weighted_Score']]
